+++ plugins/watch-for-changes/index.d.ts
@@ -1,0 +1,2 @@
+declare let _default: {};
+export default _default;